Bugfix: clamping objects used by several processflows
Issue: when loading all the processflows, the list of related clampings
of the defaultClamping is not reset/cloned. Therefor, it will be amended
everytime a new processflow gets loaded. This means that in case the
default clamping is the same for 2 or more flows, the list of related
clampings is actually the union of all the processflows having this
specific default clamping.

Note that the list of related clampings is the list with all the
selected clampings minus the default one. It is a field of the default
clamping.

Fix: make Clamping cloneable - all the fields , like name, id, ... are
the same except for the list of related clampings. This one is a new
list for each clone.